pyproject.toml
[tool.poetry]
name = "python-script"
version = "0.1.0"
description = ""
authors = ["idenrai <geniuskwon966@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
tdqm = "^0.0.1"
argparse = "^1.4.0"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.20240311"
requests = "^2.32.3"
types-requests = "^2.32.0.20240622"
types-tqdm = "^4.66.0.20240417"
logging = "^0.4.9.6"
boto3 = "^1.34.140"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
